edit role

+++ client/src/components/body/profile/Profile.js
@@ -81,7 +81,6 @@ function Profile() {
     }
     const updateInfor = () => {
         try {
-            
             axios.patch('/user/update', {
                 name: name ? name : user.name,
                 avatar: avatar ? avatar : user.avatar
@@ -125,15 +124,18 @@ function Profile() {
 
     const handleDelete = async (id) => {
         try {
-            if(window.confirm("Are you sure you want to delete this account?"))
-            {
-                setLoading(true)
-                await axios.delete(`/user/delete/${id}`, {
-                    headers: {Authorization: token}
-                })
-                setLoading(false)
-                setCallback(!callback)
+            if(user._id !== id){
+                if(window.confirm("Are you sure you want to delete this account?"))
+                {
+                    setLoading(true)
+                    await axios.delete(`/user/delete/${id}`, {
+                        headers: {Authorization: token}
+                    })
+                    setLoading(false)
+                    setCallback(!callback)
+                }
             }
+
         } catch (err) {
             setData({...data, err: err.response.data.msg, success: ''})
         }
